enable bookmark clicking during playback
