+++ Assets/__Scripts/MapEditor/UI/SongTimelineHandleController.cs
@@ -4,22 +4,11 @@ using UnityEngine.EventSystems;
 public class SongTimelineHandleController : MonoBehaviour, IPointerUpHandler, IPointerDownHandler
 {
     [SerializeField] private SongTimelineController timeline;
-    [SerializeField] private AudioTimeSyncController atsc;
-    private bool resume = false;
-
-    private void OnEnable()
-    {
-        atsc.OnPlayToggle += OnPlayToggle;
-    }
+    [SerializeField] private TimelineInputPlaybackController tipc;
 
     public void OnPointerDown(PointerEventData fucku)
     {
-        if (atsc.IsPlaying)
-        {
-            atsc.TogglePlaying();
-            resume = true;
-        }
-
+        tipc.PointerDown();
         timeline.IsClicked = true;
         timeline.TriggerUpdate();
     }
@@ -28,21 +17,6 @@ public class SongTimelineHandleController : MonoBehaviour, IPointerUpHandler, IP
     {
         timeline.TriggerUpdate();
         timeline.IsClicked = false;
-
-        if (resume && !atsc.IsPlaying)
-        {
-            atsc.TogglePlaying();
-        }
-        resume = false;
-    }
-
-    private void OnPlayToggle(bool playing)
-    {
-        resume = false;
-    }
-
-    private void OnDestroy()
-    {
-        atsc.OnPlayToggle -= OnPlayToggle;
+        tipc.PointerUp();
     }
 }
